+++ src/util/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod linked_list;
+pub(crate) mod timeout;